chore: rename api folder
