+++ .eslintignore
@@ -3,5 +3,5 @@ assets
 content
 node_modules
 storybook-static
+client-app/modules/**/api/graphql/types.ts
 client-app/core/api/graphql/**/types.ts
-client-app/modules/**/api/types.ts
