+++ client-app/core/types/addresses.ts
@@ -1,5 +1,5 @@
 import type { CartAddressType, InputAddressType, MemberAddressType, OrderAddressType } from "@/core/api/graphql/types";
-import type { QuoteAddressType } from "@/modules/quotes/api/types";
+import type { QuoteAddressType } from "@/modules/quotes/api/graphql/types";
 
 export type AnyAddressType = (
   | MemberAddressType
