+++ client-app/modules/quotes/api/graphql/fragments/currency.graphql
