+++ client-app/modules/quotes/api/graphql/fragments/money.graphql
