+++ client-app/modules/quotes/api/graphql/fragments/property.graphql
