+++ client-app/modules/quotes/api/graphql/fragments/quoteAttachment.graphql
