+++ client-app/modules/quotes/api/graphql/mutations/approveQuoteRequest/approveQuoteRequestMutation.graphql
