+++ client-app/modules/quotes/api/graphql/mutations/removeQuoteItem/removeQuoteItemMutation.graphql
