+++ client-app/modules/quotes/api/graphql/mutations/updateQuoteAttachments/index.ts
