+++ client-app/modules/quotes/api/graphql/queries/getQuote/getQuoteQuery.graphql
