+++ client-app/modules/quotes/api/graphql/queries/getQuote/index.ts
