+++ client-app/modules/quotes/api/graphql/queries/getQuotes/getQuotesQuery.graphql
