+++ client-app/modules/quotes/api/graphql/queries/getQuotes/index.ts
