+++ client-app/modules/quotes/api/graphql/types.ts
