+++ client-app/modules/quotes/pages/edit-quote.vue
@@ -139,7 +139,7 @@ import QuoteLineItems from "../components/quote-line-items.vue";
 import { useUserQuote } from "../useUserQuote";
 import type { MemberAddressType } from "@/core/api/graphql/types";
 import type { AnyAddressType } from "@/core/types";
-import type { QuoteAddressType, QuoteItemType, QuoteType } from "@/modules/quotes/api/types";
+import type { QuoteAddressType, QuoteItemType, QuoteType } from "@/modules/quotes/api/graphql/types";
 import type { StringSchema } from "yup";
 import AddOrUpdateAddressModal from "@/shared/account/components/add-or-update-address-modal.vue";
