+++ sonar-project.properties
@@ -2,7 +2,7 @@ sonar.projectKey=VirtoCommerce_vc-theme-b2b-vue
 sonar.organization=virto-commerce
 
 # General code rules
-sonar.exclusions=client-app/core/api/graphql/**/types.ts, client-app/modules/**/api/types.ts
+sonar.exclusions=client-app/core/api/graphql/**/types.ts, client-app/modules/**/api/graphql/types.ts
 
 # Code duplication rules
 sonar.cpd.exclusions=**/*.mock.ts, **/*.test.ts
